requirements.txt
deep-translator==1.11.4
requests==2.31.0
google-generativeai==0.3.2

